# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Sky Journal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /media/read-urls:
    post:
      operationId: resolveMediaReadUrls
      tags: [stories]
      summary: Resolve managed media paths into owner-authorized read URLs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaReadUrlsInput"
      responses:
        "200":
          description: Canonical media paths mapped to raw public paths or short-lived owner URLs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaReadUrlsResponse"
        "400":
          description: Invalid request
        "401":
          description: Authentication required
        "404":
          description: Media unavailable
...
components:
  schemas:
    MediaReadUrlsInput:
      type: object
      additionalProperties: false
      properties:
        paths:
          type: array
          maxItems: 100
          items:
            type: string
            minLength: 1
            maxLength: 2048
      required: [paths]
    MediaReadUrlsResponse:
      type: object
      properties:
        urls:
          type: object
          additionalProperties:
            type: string
      required: [urls]
...