# lib/api-zod/src/generated/types/mediaReadUrlsInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sky Journal API
 * OpenAPI spec version: 0.2.0
 */

export interface MediaReadUrlsInput {
  /** @maxItems 100 */
  paths: string[];
}