# lib/api-zod/src/generated/types/mediaReadUrlsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sky Journal API
 * OpenAPI spec version: 0.2.0
 */
import type { MediaReadUrlsResponseUrls } from "./mediaReadUrlsResponseUrls";

export interface MediaReadUrlsResponse {
  urls: MediaReadUrlsResponseUrls;
}